# Conflicts:
#	2-11-16 9_00PM/Assets/Scripts/PointConverter.cs
#	2-11-16 9_00PM/Library/CurrentLayout.dwlt
#	2-11-16 9_00PM/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll
#	2-11-16 9_00PM/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll.mdb
#	2-11-16 9_00PM/Library/ScriptAssemblies/Assembly-CSharp.dll
#	2-11-16 9_00PM/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	2-11-16 9_00PM/Library/UnityAssemblies/UnityEditor.Advertisements.dll
#	2-11-16
9_00PM/Library/UnityAssemblies/UnityEditor.EditorTestsRunner.dll
#	2-11-16 9_00PM/Library/UnityAssemblies/UnityEditor.Graphs.dll
#	2-11-16
9_00PM/Library/UnityAssemblies/UnityEditor.LinuxStandalone.Extensions.dll
#	2-11-16 9_00PM/Library/UnityAssemblies/UnityEditor.Networking.dll
#	2-11-16 9_00PM/Library/UnityAssemblies/UnityEditor.TreeEditor.dll
#	2-11-16 9_00PM/Library/UnityAssemblies/UnityEditor.UI.dll
#	2-11-16
9_00PM/Library/UnityAssemblies/UnityEditor.WindowsStandalone.Extensions.dll
#	2-11-16 9_00PM/Library/assetDatabase3
#	2-11-16 9_00PM/Library/expandedItems
#	2-11-16 9_00PM/Library/metadata/00/00000000000000004000000000000000
#	2-11-16 9_00PM/Library/metadata/00/00000000000000006100000000000000
#	2-11-16 9_00PM/Library/metadata/38/38efa3f36247c2c429af78f3d7dc2f6b
#	2-11-16 9_00PM/Library/metadata/5d/5d1afed1afe4b774987df5df5e4e1211
#	2-11-16 9_00PM/Library/metadata/88/8843f3202f3f43540b3ba36e4c669f1f
#	2-11-16 9_00PM/Library/metadata/d0/d01fe6fbeefe63d41858f87509e50325

+++ 2-11-16 9_00PM/Assets/Scripts/DataCollector.cs
@@ -3,12 +3,12 @@ using System.Collections;
 
 public class DataCollector : MonoBehaviour {
     [SerializeField]
-    private float radPoints = 0;
-    public float getRadPoints()
+    private double radPoints = 0;
+    public double getRadPoints()
     {
         return this.radPoints;
     }
-    public void setRadPoints(float rad)
+    public void setRadPoints(double rad)
     {
         this.radPoints = rad;
     }
@@ -35,13 +35,13 @@ public class DataCollector : MonoBehaviour {
     }
     [SerializeField]
 
-    private float pointsPerSec = 0;
+    private double pointsPerSec = 0;
 
-    public float getPointPerSec()
+    public double getPointPerSec()
     {
         return this.pointsPerSec;
     }
-    public void setPointsPerSec(float points)
+    public void setPointsPerSec(double points)
     {
         this.pointsPerSec = points;
     }
